Improve naming, add optionals to WeatherData
Change the naming of some files and namespaces to better
reflect functionality.

Change WeatherData data members to be optionals, to
better handle cases where those data are missing or
unavailable.

Modify unit tests and other files based on the above changes.

Create a driver class for the parseweather cli script and
start creating the script.